refactored number range filter
